requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
psycopg2-binary
python-dotenv
pyjwt
python-jose[cryptography]
pydantic-settings
pydantic[email]
requests
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==4.0.1